patients can upload and update one image of their eczema each day
