+++ src/config/constants.js
@@ -1,7 +1,7 @@
 require("dotenv").config();
 
 // console.log("WHAT IS ENV", process.env.API_URL);
-
+export const cloudName = "derma-app";
 export const apiUrl = "http://localhost:4000";
 //or  process.env.API_URL once I have figured this out
 export const DEFAULT_MESSAGE_TIMEOUT = 3000;
